requirements.txt
Flask==3.1.2
flask-cors==4.0.0
flasgger==0.9.7.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
pdfplumber==0.11.4
pytesseract==0.3.13
pillow==10.4.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==0.6.23
PyMuPDF>=1.24.0
opencv-python-headless==4.8.1.78
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.17
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
langchain-experimental==0.3.3
pgvector==0.2.5
numpy==1.26.4
pandas==2.2.3
pycryptodome==3.20.0
tabulate==0.9.0
langdetect==1.0.9
deep-translator==1.11.4
googletrans==4.0.2
ddgs==9.6.1
beautifulsoup4==4.12.3
playwright==1.49.0
SQLAlchemy==2.0.25
Werkzeug==3.1.3
